refactor(webhook_service/trigger_webhook): refactor the code to match coding guideline (SCRUM-244)
